Add webhook integration for deploying latest release
Add a webhook which allows deploying a release for a given project. The
use case for this is having external services that aren't CI servers
being able to trigger deploys. An example is that we have a button in
the office which can be physically pushed that will ping the endpoint to
perform a deploy.