Only show buddy confirmationsettings when the feature is enabled
The 'Does this stage impact production servers?' and 'Confirm before
deployment' options for a stage are only ever used to confirm
deployments when the buddy feature is activated. Therefore we do not
need to display them when the feature is disabled. Simiarly to how
fields are hidden when the 'deploy group feature' is not enabled.