Fix bug where some buildkite integrations would not create a relase
Buildkite sometimes sends `build.finished` events rather than `build`
and in those cases we weren't creating a release. I have removed that
check because we don't actually need to check the event type, because we
can get enough info from the fact that we have a successful build with a
valid commit.