Fix loading commit when comparing to master
The changeset is cached with cache key is based on previous commit
and the tag when deploying. If the tag is a branch name, such as
`master`, the cache key is same leads to us seeing old data. If we
push more commit to the branch, new changeset won't be updated.