Adding tests for the Create Kubernetes Release wizard.
Raising an exception if no DeployGroups or Roles have been provided, as no ReleaseDocs will be created.

Fixing the use of the ArgumentError class (tests were interpreting this as a method call)

Moving the responsibility of parsing the kubernetes releases objects to the service.
Adding error handling.

Fixing karma config.

The Bootstrap UI directive is, apparently, afecting the current flag. Just making sure that the flags are correctly set independently of the behaviour of the directive.

Parsing the builds in the build service instead of the controller.