Added extra metadata to the Deployment resource. Some fixes for the DeployWatcher (not finished). Some fixes to the Dashboard.
DeployWatcher is now listening to updates coming from any release affected by a new Deployment.

Making the DeployWatcher supervised to protect it from unexpected crashes.

Addressing code review comments.
Sending SSE event with the up-to-date current number of failed pods each time we get a cluster event with a pod error.

Fixing bugs: after re-importing the roles, we were getting a few errors due to the soft-deletion of the previous roles; the Deployment resource selector cannot use the Role ID, as it will be unable to select the old RCs when the roles are re-imported.