RFC: Reduce docker image size
This reduces the docker image size from 1GB to 790MB, 23% smaller.
Further changes may be possible, but they are probably more intrusive.
These are pretty simple and yet the shrink is considerable. So this
feels good enough for a start, at least.

This is useful for us because our internet connection is pretty slow.