Fix typo
Commit e3594b032196991b3c076545ecd03838ca2d5b8b introduced a
push_to_latest param in push_image() callers, but when defining it in
push_image() it was called "tag_as_latest".

This is clearly a typo, and without having any preference, I choose to
make the param in push_image() "push_to_latest". Please feel free to
change it, just let's make sure it's used with the same name :)