Allow to manually deploy branches with / in it's name
When doing a manual build from a branch name with a slash, for example
"feature/kubernetes", the deploy of that build fails with this error:

  *********** Couldn't deploy: Deployment.extensions "wido-web" is
  invalid: [metadata.labels: Invalid value: "feature/kubernetes": must
  have at most 63 characters, matching regex
  (([A-Za-z0-9][-A-Za-z0-9_.]*)?[A-Za-z0-9])?: e.g. "MyValue" or "",

Using the git_sha workarounds this, although when a tag is used, then
the hash is showed too.

Maybe the field should be renamed if the sha is what we want. But I did
this small patch just to get samson working on my setup and wanted to
know what you think about this.