Fix Dockerfile
This reverts the nodejs parts of commit 4a946a19 that changed the Dockerfile.
The bug hit it is this: npm/npm#9863 (although reported to an older version, it
still applies and bug reports to newer versions of npm where closed as
duplicates of this bug).

This fixes the build that is not working since that change and docker hub fails
to build images automatically since then.