Avoid DataDog spam
Sending events to DataDog which have an `@` followed by something vaguely email shaped will cause DataDog to send an email to that email account.

On the surface, this isnt so bad, but when you use application specific email addresses you are left in a situation where you are required to create an entirely new DataDog account just to unsubscribe that email address from said spam. 

This, apparently, is a feature.

This change should still allow notifications to be sent, but more importantly, only if it matches a known DataDog account, allowing the individual to manage their spam preferences accordingly.