Always redirect to login path when unauthorized
Not sure why it was set to redirect back to the referrer at the first
place, but since Ledger contains links to Samson, and if the user is not
logged into Samson, it redirects the user back to Ledger which is
confusing. I think we should always redirect the user to login page if
unauthorized.