stream logs from live/dead pods to fight incomplete logs without having to sleep
We had many deploys where asking for the logs even 2-10s after the pod died
did not result in the complete logs.

FYI using no container argument was often more reliable at getting logs.

With this approach we can increase the timeout without affecting good deploys
instead of always sleeping which would be ineffective.
We also get notified when the logs did not finish so we know to inspect manually.