Change status message for successful deploy
The current wording:

Alice (with Bob) deployed v11105 to Production deploy 8 minutes ago, it took 5
minutes 36 seconds.

makes it hard to tell if the deploy was started or finished 8 minutes ago. This
commit changes the wording to be:

Alice (with Bob) started deploying v11105 to Production deploy 8 minutes ago,
it took 5 minutes 36 seconds.