[TOOLS-2157] fix thread contention issue during docker builds
A thread in the Kubernetes::BuildFinder class was calling #reload on an
instance of build that was being used by the DockerBuilderService class,
and the state of that object was being discarded. This fix safely reloads
the data from the DB without mutating the object, at least while it's
in an active state.