Added resolve endpoint to secrets controller
This allows resolving secrets to get their paths in the context of a
project and deploy group.

This could be used for external integrations, while keeping the logic of
resolving secrets, and the currently underlying path structure in place,
without needing to explicitly know where a key resides in vault.

ref: https://zendesk.atlassian.net/browse/GUIDEOPS-1284