Adress feedback from @grosser
We now support both project_permalink, and project_id combinations in
query params.

We properly support wildcard expanded secrets.

Any secret which is not resolved will have a value of "null" in the
resolved hash.

Keys can be given as a comma delimited list in the query params as well
as a list.