Fix regression causing prerequisite stages to stop working
SamsonPrerequisisteStages::validate_deployed_to_all_prerequisiste_stages
was changed to accept a commit instead of a reference, but it was still
being called with a reference from the before_deploy hook. This made
prerequisiste stages never be satisfied (unless the deploy was done with
the acutal commit SHA).

Fix it by resolving the ref to a commit in the before_deploy hook.