Adding permalink to deploy csv_export
Adding a column (permalink) to the export CSV functionality. I need for automation the SOC2 control C.24. In my automation, i would scanning through all the deploys and need the permalink for it

Previously one could use any deploy number in the URL (regardless of the project parameter) to fetch the deploy page. I think this was fixed recently.