feat(changeset): change how recover Pull Request ID
As we could add the Pull Request ID in any place of the commit message,
we would like to change how we parse the commit message. The only
mandatory thing is to have parenthesis around it ex: (#12347).
We also remove the commit message truncation to avoid ignoring PR number
if located more than 80 characters away (at the end of the first line or
in the next lines)